eliminar api.js duplicado
